CartDetail còn quantity

+++ src/java/Controller/CartDetails.java
@@ -64,12 +64,11 @@ public class CartDetails extends HttpServlet {
          Customer cus = (Customer) session.getAttribute("cus");
          DAOCart db = new DAOCart();
          DAOProduct db1 = new DAOProduct();
-         DAOCategoryProduct db2 = new DAOCategoryProduct();
-    
-          session.setAttribute("cart", db.getListCart(cus.getCustomerID()));
-         request.setAttribute("Cate1", db2.getCategoryProductProduct());
+         DAOCategoryProduct db2 = new DAOCategoryProduct();    
+        session.setAttribute("cart", db.getListCart(cus.getCustomerID()));
+        request.setAttribute("Cate1", db2.getCategoryProductProduct());
         request.setAttribute("CategoryB", db1.ListCatogoryAndBrand());
-         request.setAttribute("list", db.getListCart(cus.getCustomerID()));
+        request.setAttribute("list", db.getListCart(cus.getCustomerID()));
         request.getRequestDispatcher("Views/CartDetails.jsp").forward(request, response);
     }
 
@@ -87,13 +86,15 @@ public class CartDetails extends HttpServlet {
        String Quantity = request.getParameter("quantity");
        String CardID = request.getParameter("cartid");
        String pid = request.getParameter("pid");
+       String delete = request.getParameter("delete");
        DAOCart db = new DAOCart();
        if(Quantity !=null && CardID !=null){
+        
        db.UpdateCartQuantity(Integer.parseInt(CardID), Integer.parseInt(Quantity),Integer.parseInt(pid));
        }
       String CartItem = request.getParameter("cartitemid");
       
-       if(CardID != null && CartItem !=null){
+       if(CardID != null && CartItem !=null && delete != null){
            db.DeleteCardItems(Integer.parseInt(CardID), Integer.parseInt(CartItem));
        }
        request.setAttribute("scroll", "scroll");
